backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
pydantic
passlib[argon2]
argon2-cffi
PyJWT
email-validator